+++ crates/entrypoint/src/cli/command/azure/azure_command.rs
@@ -1,6 +1,7 @@
 use super::group::AzureGroupArgs;
 use super::pim::AzurePimArgs;
 use super::policy::AzurePolicyArgs;
+use super::role::AzureRoleArgs;
 use super::tag::AzureTagArgs;
 use crate::noninteractive::prelude::audit_azure;
 use clap::Subcommand;
@@ -17,6 +18,8 @@ pub enum AzureCommand {
     Policy(AzurePolicyArgs),
     /// Manage Azure resource tags.
     Tag(AzureTagArgs),
+    /// Manage Azure role-based access control.
+    Role(AzureRoleArgs),
     /// Manage Azure Privileged Identity Management operations.
     Pim(AzurePimArgs),
 }
@@ -34,6 +37,9 @@ impl AzureCommand {
                 args.invoke().await?;
             }
             AzureCommand::Tag(args) => {
+                args.invoke().await?;
+            }
+            AzureCommand::Role(args) => {
                 args.invoke().await?;
             }
             AzureCommand::Pim(args) => {